requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
billiard==4.2.2
celery==5.3.6
certifi==2025.8.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
Django==5.0.2
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kombu==5.5.4
mysqlclient==2.2.4
openai==1.12.0
packaging==25.0
pillow==10.2.0
prompt_toolkit==3.0.52
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.0.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.6.0
